Fifo task queue sheduler
FifoTaskQueue is a FIFO task queue for .Net Framework 4.7.2. and it is able
to interact with the GUI controlls or to process Task at the backend.

The primary goal of this component is to allow asynchronous tasks sequentially.

It also observes the status of the processing Task for a tracking overview and
supports task cancellation explicitly or after a given elapsed time.
Once a task has been cancelled of failed, these subordinated tasks will be
canceled before starting setting the TaskStatus of each task properly.